requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu126
torch==2.7.0+cu126
torchvision
torchaudio


einops>=0.7
numpy>=1.23
scipy>=1.9
matplotlib>=3.6
seaborn>=0.12
tqdm>=4.64
pyyaml>=6.0
pandas>=1.5
scikit-learn>=1.2
zarr>=2.16
numcodecs>=0.12